+++ P_260729/Plugins/JMMonsterFramework/Source/JMMonsterFrameworkRuntime/JMMonsterFrameworkRuntime.Build.cs
@@ -12,6 +12,7 @@ public class JMMonsterFrameworkRuntime : ModuleRules
             "CoreUObject",
             "Engine",
             "AIModule",
+            "NavigationSystem",
             "StateTreeModule",
             "GameplayStateTreeModule"
         });